26. Remove Duplicates from Sorted Array
