+++ src/test/kotlin/me/bossm0n5t3r/leetcode/removeDuplicatesFromSortedArray/RemoveDuplicatesFromSortedArrayTest.kt
@@ -1,10 +1,10 @@
-package removeDuplicatesFromSortedArray
+package me.bossm0n5t3r.leetcode.removeDuplicatesFromSortedArray
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class RemoveDuplicatesFromSortedArrayTest {
-    private val removeDuplicatesFromSortedArray = RemoveDuplicatesFromSortedArray()
+    private val removeDuplicatesFromSortedArray = RemoveDuplicatesFromSortedArray.Solution()
 
     data class RemoveDuplicatesFromSortedArrayTestData(
         val nums: IntArray,
